Issue-366: Failing to actually a poll device that returns "NoSuchObject" for sysDecr or sysUpTime (#367)
If sysDescr or sysUpTime return "NoSuchObject" then accept the return
and process normally, otherwise clear it and change status to an error.
The behavior changed due to the patch for #272, and I have not been able
to poll a few devices with spine newer than 1.2.21. I have tested this
change on 1.2.27 (my current version in production)

Co-authored-by: TheWitness <thewitness@cacti.net>